+++ src/plugin.cpp
@@ -59,10 +59,12 @@ void init(Plugin* p) {
 	p->addModel(modelSickoSampler2);
 	p->addModel(modelSimpleSeq4);
 	p->addModel(modelSlewer);
+	p->addModel(modelStepSeq);
 	p->addModel(modelSwitcher);
 	p->addModel(modelSwitcherSt);
 	p->addModel(modelSwitcher8);
 	p->addModel(modelToggler);
 	p->addModel(modelTogglerCompact);
+	p->addModel(modelTrigSeq);
 	p->addModel(modelWavetabler);
 }